Init: leftProfileCard component init #112
