+++ frontend/src/components/Profile/Left/index.js
@@ -1,0 +1,1 @@
+export {default as MyProfileCard} from "./myProfileCard.js";
