+++ frontend/src/pages/profilePage.js
@@ -1,21 +1,26 @@
 import { Component, createComponent } from "../core/index.js";
 import { BackIconButton } from "../components/Button/index.js";
+import { MyProfileCard } from "../components/Profile/left/index.js";
 import Store from "../store/index.js";
 
 class ProfilePage extends Component {
   render() {
     this.container = document.createElement("div");
-    this.container.className = "flex items-center justify-center min-h-screen bg-no-repeat bg-cover";
+    this.container.className =
+      "flex items-center justify-center min-h-screen bg-no-repeat bg-cover";
     if (Store.state.darkMode) {
-      this.container.style.backgroundImage = 'url("/static/assets/images/net-dark.svg")';
+      this.container.style.backgroundImage =
+        'url("/static/assets/images/net-dark.svg")';
     } else {
-      this.container.style.backgroundImage = 'url("/static/assets/images/net.svg")';
+      this.container.style.backgroundImage =
+        'url("/static/assets/images/net.svg")';
     }
     this.container.style.backgroundPosition = "calc(100% - 680px) center";
 
     // grid 레이아웃을 사용하는 컨테이너 생성
     const gridContainer = document.createElement("div");
-    gridContainer.className = "grid grid-cols-12 gap-10 flex flex-col items-center justify-center min-h-screen";
+    gridContainer.className =
+      "grid grid-cols-12 gap-10 flex flex-col items-center justify-center min-h-screen";
 
     // BackIconButton 추가
     const backIcon = createComponent(BackIconButton, {});
@@ -23,29 +28,16 @@ class ProfilePage extends Component {
     gridContainer.appendChild(backIcon);
 
     // leftProfileCard 추가
-    const leftProfileCard = document.createElement("div");
-    leftProfileCard.className = "col-start-6 col-span-3 relative w-[296px] h-[632px]";
-
-    const board = document.createElement("div");
-    board.className = "w-[296px] h-[632px] relative";
-
-    const lowerBoard = document.createElement("div");
-    lowerBoard.className = "w-[296px] h-[632px] left-0 top-0 absolute bg-primary-card_background rounded-2xl";
-
-    const upperBoard = document.createElement("div");
-    upperBoard.className = "w-[296px] h-[280px] left-0 top-0 absolute bg-primary-button rounded-tl-2xl rounded-tr-2xl";
-
-    board.appendChild(lowerBoard);
-    board.appendChild(upperBoard);
-
-    leftProfileCard.appendChild(board);
+    const leftProfileCard = createComponent(MyProfileCard, {});
 
     // rightProfileCard 추가
     const rightProfileCard = document.createElement("div");
-    rightProfileCard.className = "col-start-9 col-span-4 relative w-[408px] h-[632px]";
+    rightProfileCard.className =
+      "col-start-9 col-span-4 relative w-[408px] h-[632px]";
 
     const board2 = document.createElement("div");
-    board2.className = "w-[408px] h-[632px] left-0 top-0 absolute bg-primary-card_background rounded-2xl border-8 border-primary-button";
+    board2.className =
+      "w-[408px] h-[632px] left-0 top-0 absolute bg-primary-card_background rounded-2xl border-8 border-primary-button";
 
     rightProfileCard.appendChild(board2);
